AC-493 feature: put duplicate code into component and fixed typo issue in constants
